changed serialization to be more consistent with the theta sketches

+++ src/test/java/com/yahoo/sketches/tuple/HeapArrayOfDoublesQuickSelectSketchTest.java
@@ -170,7 +170,7 @@ public class HeapArrayOfDoublesQuickSelectSketchTest {
     for (int i = 0; i < numberOfUniques; i++) sketch1.update(i, new double[] {1.0});
     UpdatableArrayOfDoublesSketch sketch2 = new HeapArrayOfDoublesQuickSelectSketch(new NativeMemory(sketch1.toByteArray()));
     Assert.assertTrue(sketch2.isEstimationMode());
-    Assert.assertEquals(sketch2.getEstimate() / (double) numberOfUniques, 1.0, 0.01);
+    Assert.assertEquals(sketch2.getEstimate() / numberOfUniques, 1.0, 0.01);
     Assert.assertEquals(sketch2.getRetainedEntries() / (double) numberOfUniques, 0.5, 0.01);
     Assert.assertEquals(sketch1.getTheta(), sketch2.getTheta());
   }
